+++ dwds/CHANGELOG.md
@@ -9,6 +9,7 @@
   - Fix crash when using flutter tools with web server device.
   - Remove clearing all scripts on page load for extension debugger.
 - Fix breakpoints not hitting after changing a base in index.html.
+- Find best locations for call frames, breakpoints, or expression evaluation.
 
 **Breaking changes:**
 - Add `basePath` parameter to `FrontendServerRequireStrategy`.